fix(trellis): bind immutable serving launch state
